backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
openai==1.99.3
pinecone-client==3.0.0
tiktoken==0.5.1
pydantic==2.5.0
python-multipart==0.0.6